bug fixes + refactoring

+++ srcuda/material.cuh
@@ -109,7 +109,7 @@ private:
     // schlick approximation for refl
     float r0 = (1 - refractive_index) / (1 + refractive_index);
     r0 = r0 * r0;
-    return r0 + (1 - r0) * std::pow((1 - cos_theta), 5);
+    return r0 + (1 - r0) * powf((1 - cos_theta), 5);
   }
 };
